+++ src/ircd.c
@@ -282,7 +282,9 @@ set_time(void)
  if (newtime < CurrentTime)
  {
   ircsprintf(to_send, "System clock is running backwards - (%lu < %lu)",
-             newtime, CurrentTime);
+		(unsigned long) newtime,
+		(unsigned long) CurrentTime);
+
   report_error(to_send, me.name, 0);
   set_back_events(CurrentTime - newtime);
  }
